requirements.txt
requests==2.26.0
python-dotenv==0.19.1
# Flask==2.0.2
Flask==2.3.2
Werkzeug>=2.3.3
matplotlib==3.9.2